For #21611: Show the jump back in Contextual Hints independently of the home onboarding dialog.
