+++ app/src/androidTest/java/org/mozilla/fenix/ui/CollectionTest.kt
@@ -40,7 +40,7 @@ class CollectionTest {
 
     @Before
     fun setUp() {
-        activityTestRule.activity.applicationContext.settings().hasShownHomeOnboardingDialog = true
+        activityTestRule.activity.applicationContext.settings().shouldShowJumpBackInCFR = false
 
         mockWebServer = MockWebServer().apply {
             dispatcher = AndroidAssetDispatcher()
