+++ app/src/androidTest/java/org/mozilla/fenix/ui/HomeScreenTest.kt
@@ -119,7 +119,7 @@ class HomeScreenTest {
     @Test
     fun dismissOnboardingUsingHelpTest() {
         val settings = activityTestRule.activity.applicationContext.settings()
-        settings.hasShownHomeOnboardingDialog = true
+        settings.shouldShowJumpBackInCFR = false
         homeScreen {
             verifyWelcomeHeader()
         }.openThreeDotMenu {
