+++ app/src/androidTest/java/org/mozilla/fenix/ui/NavigationToolbarTest.kt
@@ -12,6 +12,7 @@ import org.junit.Before
 import org.junit.Ignore
 import org.junit.Rule
 import org.junit.Test
+import org.mozilla.fenix.ext.settings
 import org.mozilla.fenix.helpers.AndroidAssetDispatcher
 import org.mozilla.fenix.helpers.HomeActivityTestRule
 import org.mozilla.fenix.helpers.TestAssetHelper
@@ -41,6 +42,8 @@ class NavigationToolbarTest {
             dispatcher = AndroidAssetDispatcher()
             start()
         }
+        val settings = activityTestRule.activity.settings()
+        settings.shouldShowJumpBackInCFR = false
     }
 
     @After
