+++ app/src/androidTest/java/org/mozilla/fenix/ui/TabbedBrowsingTest.kt
@@ -48,14 +48,11 @@ class TabbedBrowsingTest {
 
     @Before
     fun setUp() {
-        activityTestRule.activity.applicationContext.settings().hasShownHomeOnboardingDialog = true
+        activityTestRule.activity.applicationContext.settings().shouldShowJumpBackInCFR = false
         mockWebServer = MockWebServer().apply {
             dispatcher = AndroidAssetDispatcher()
             start()
         }
-
-        val settings = activityTestRule.activity.applicationContext.settings()
-        settings.hasShownHomeOnboardingDialog = true
     }
 
     @After
