+++ app/src/debug/res/raw/initial_experiments.json
@@ -12,7 +12,7 @@
                       "value": {
                            "sections-enabled": {
                                "topSites": true,
-                               "recentExplorations": false,
+                               "recentExplorations": true,
                                "recentlySaved": false,
                                "jumpBackIn": false,
                                "pocket": false
